Finalize merge
