+++ src/protocols/tcp.rs
@@ -58,7 +58,6 @@ impl Protocol for TcpHeader {
     /// Returns the name list of all field of the protocols.
     ///
     /// Header names are suffixed with an index (e.g., `tcp_sprt_0`, `tcp_sprt_1`).
-
     fn get_headers_name() -> Vec<String> {
         let fields = vec![
             ("tcp_sprt", 16),
@@ -171,7 +170,7 @@ mod tcp_header_tests {
     }
 
     #[test]
-    fn test_tcp_header_get_headers() {
+    fn test_tcp_header_get_headers_name() {
         let expected_headers = vec![
             "tcp_sprt_0",
             "tcp_sprt_1",
@@ -655,7 +654,7 @@ mod tcp_header_tests {
             "tcp_opt_319",
         ];
 
-        let headers = TcpHeader::get_headers();
+        let headers = TcpHeader::get_headers_name();
         for (i, expected) in expected_headers.iter().enumerate() {
             assert_eq!(
                 headers[i], *expected,
